feat: V2Controller 분리
